chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
qdrant-client==1.16.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
